requirements.txt
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
colorclass==2.2.2
colorlog==6.7.0
croniter==1.3.0
faust==1.10.4
frozenlist==1.3.3
idna==3.4
kafka-python==1.4.6
mode==4.3.2
multidict==6.0.4
mypy-extensions==0.4.4
opentracing==1.3.0
python-dateutil==2.8.2
robinhood-aiokafka==1.1.6
six==1.16.0
terminaltables==3.1.10
venusian==1.2.0
yarl==1.9.2
